sonar bugfix

+++ eservice_adm/src/main/java/com/twfhclife/adm/service/impl/FuncMgntServiceImpl.java
@@ -25,19 +25,19 @@ import java.util.List;
  */
 @Service
 public class FuncMgntServiceImpl implements IFuncMgntService {
-	
+
 	@Autowired
 	private FunctionItemDao functionItemDao;
-	
+
 	@Autowired
 	private FunctionDivDao functionDivDao;
-	
+
 	@Autowired
 	FuncMgmtClient funcMgmtClient;
-	
+
 	/**
 	 * 根據系統別取得所有功能.
-	 * 
+	 *
 	 * @return 回傳該系統的所有功能.
 	 */
 	@RequestLog
@@ -61,14 +61,14 @@ public class FuncMgntServiceImpl implements IFuncMgntService {
 					funcItemVo.setDivArr(StringUtils.join(divNames, ','));
 				}
 			}
-		}		
-		
+		}
+
 		return funcItemList;
 	}
 
 	/**
 	 * 判斷功能名稱是否存在.
-	 * 
+	 *
 	 * @param functionVo FunctionItemVo
 	 * @return 回傳功能名稱是否存在
 	 */
@@ -89,10 +89,10 @@ public class FuncMgntServiceImpl implements IFuncMgntService {
 		}
 		return funNameExist;
 	}
-	
+
 	/**
 	 * 新增功能節點.
-	 * 
+	 *
 	 * @param functionVo FunctionItemVo
 	 * @return 回傳影響筆數
 	 */
@@ -113,41 +113,43 @@ public class FuncMgntServiceImpl implements IFuncMgntService {
 		}
 		return result;
 	}
-	
+
 	/**
 	 * 新增功能節點.
-	 * 
+	 *
 	 * @param functionVo FunctionItemVo
 	 * @return 回傳影響筆數
 	 */
 	@RequestLog
 	@Transactional
 	public int addFunctionItem(FunctionItemVo functionVo) {
-		
+
 		if(functionVo!=null) {
 			if(functionVo.getSort()==null || "".equals(functionVo.getSort().trim())) {
 				/**
 				 * fixed sqlserver.jdbc.SQLServerException
-				 * 將資料類型從nvarchar轉換到numberic時發生錯誤 
+				 * 將資料類型從nvarchar轉換到numberic時發生錯誤
 				 */
 				functionVo.setSort("0");//fixed sqlserver.jdbc.SQLServerException
 			}
 		}
-		
+
 		int result = functionItemDao.addFunctionItem(functionVo);
 		if (result > 0) {
-			if (functionVo.getFunctionId() != null) {
+			if (functionVo != null && functionVo.getFunctionId() != null) {
 				functionDivDao.deleteByFunId(functionVo.getFunctionId());
 			}
-			
-			String[] divArr = functionVo.getDivArr().split(",");
-			for (String divName : divArr) {
-				if (MyStringUtil.isNotNullOrEmpty(divName)) {
-					FunctionDivVo functionDivVo = new FunctionDivVo();
-					functionDivVo.setFunctionId(new BigDecimal(functionVo.getFunctionId()));
-					functionDivVo.setDivName(divName);
-					functionDivVo.setCreateUser(functionVo.getCreateUser());
-					functionDivDao.insertFunctionDiv(functionDivVo);
+
+			if (functionVo != null) {
+				String[] divArr = functionVo.getDivArr().split(",");
+				for (String divName : divArr) {
+					if (MyStringUtil.isNotNullOrEmpty(divName)) {
+						FunctionDivVo functionDivVo = new FunctionDivVo();
+						functionDivVo.setFunctionId(new BigDecimal(functionVo.getFunctionId()));
+						functionDivVo.setDivName(divName);
+						functionDivVo.setCreateUser(functionVo.getCreateUser());
+						functionDivDao.insertFunctionDiv(functionDivVo);
+					}
 				}
 			}
 		}
@@ -156,35 +158,39 @@ public class FuncMgntServiceImpl implements IFuncMgntService {
 
 	/**
 	 * 更新功能節點.
-	 * 
+	 *
 	 * @param functionVo FunctionItemVo
 	 * @return 回傳影響筆數
 	 */
 	@RequestLog
 	@Transactional
 	public int updateFunctionItem(FunctionItemVo functionVo) {
-		
+
 		if(functionVo!=null) {
 			if(functionVo.getSort()==null || "".equals(functionVo.getSort().trim())) {
 				/**
 				 * fixed sqlserver.jdbc.SQLServerException
-				 * 將資料類型從nvarchar轉換到numberic時發生錯誤 
+				 * 將資料類型從nvarchar轉換到numberic時發生錯誤
 				 */
 				functionVo.setSort("0");//fixed sqlserver.jdbc.SQLServerException
 			}
 		}
-		
+
 		int result = functionItemDao.updateFunctionItem(functionVo);
 		if (result > 0) {
-			functionDivDao.deleteByFunId(functionVo.getFunctionId());
-			String[] divArr = functionVo.getDivArr().split(",");
-			for (String divName : divArr) {
-				if (MyStringUtil.isNotNullOrEmpty(divName)) {
-					FunctionDivVo functionDivVo = new FunctionDivVo();
-					functionDivVo.setFunctionId(new BigDecimal(functionVo.getFunctionId()));
-					functionDivVo.setDivName(divName);
-					functionDivVo.setCreateUser(functionVo.getCreateUser());
-					functionDivDao.insertFunctionDiv(functionDivVo);
+			if (functionVo != null && functionVo.getFunctionId() != null) {
+				functionDivDao.deleteByFunId(functionVo.getFunctionId());
+			}
+			if (functionVo != null) {
+				String[] divArr = functionVo.getDivArr().split(",");
+				for (String divName : divArr) {
+					if (MyStringUtil.isNotNullOrEmpty(divName)) {
+						FunctionDivVo functionDivVo = new FunctionDivVo();
+						functionDivVo.setFunctionId(new BigDecimal(functionVo.getFunctionId()));
+						functionDivVo.setDivName(divName);
+						functionDivVo.setCreateUser(functionVo.getCreateUser());
+						functionDivDao.insertFunctionDiv(functionDivVo);
+					}
 				}
 			}
 		}
@@ -193,7 +199,7 @@ public class FuncMgntServiceImpl implements IFuncMgntService {
 
 	/**
 	 * 刪除功能節點.
-	 * 
+	 *
 	 * @param functionVo FunctionItemVo
 	 * @return 回傳影響筆數
 	 */
